EOM
It added EOM.pdf It has my calculations.
In the data.py I used a modified version that still uses units in which
l=1,g=1 but the simplification m1=m2 no longer is made.
"eps" is the mass of the outer ball in units of the mass of the inner.